requirements.txt
robotframework==7.2.2
robotframework-seleniumlibrary==6.7.0
selenium==4.29.0
chromedriver-autoinstaller==0.2.2